Updated documentation

+++ .github/generate_docs.bat
@@ -12,7 +12,7 @@ echo Documentation generated successfully.
 robocopy "_build\html" "html" /E /MOVE /R:3 /W:5 /V
 echo HTML files moved successfully(? - throws error dont know why when using || exit /b 1).
 
-rmdir /S /Q "_build\html" || (echo Failed to delete html folder inside _build. Exiting.. & exit/b 1)
+IF EXIST "_build\html" rmdir /S /Q "_build\html" || (echo Failed to delete html folder inside _build. Exiting.. & exit/b 1)
 echo Deleted html files inside _build
 
 git commit -a -m "Updated documentation" || (echo Failed to commit changes. Exiting... & exit /b 1)
